Optimize existing solution to leverage schedule

+++ jest.setup.ts
@@ -1,19 +1,26 @@
 import { execSync } from 'child_process';
-import { expect, jest, beforeAll, describe, it, beforeEach,afterAll } from "@jest/globals";
+import { expect, jest, beforeAll, describe, it, beforeEach,afterAll,afterEach } from "@jest/globals";
 import connection from "./src/config/database";
 import {server} from './src/index'
+import FreightSchedule from "./src/models/freightSchedule";
+import Schedule from "./src/models/schedule";
+
 // Run migrations before all tests
 beforeAll(async () => {
   process.env.NODE_ENV = 'test';
-  execSync('NODE_ENV=test npm run db:migrate', { stdio: 'inherit' });
+  execSync('NODE_ENV=test npm run db:migrate && npm run db:seed', { stdio: 'inherit' });
 });
 
 // Wipe all tables before each test
 beforeEach(async () => {
   await connection.transaction(async (transaction) => {
     const models = connection.models;
-    for (const model of Object.values(models)) {
-      await model.destroy({ where: {}, force: true, transaction });
+    const excludedTables = ["Schedule"];
+
+    for (const [modelName, model] of Object.entries(models)) {
+      if (!excludedTables.includes(modelName)) {
+        await model.destroy({ where: {}, force: true, transaction });
+      }
     }
   });
 });
